+++ src/plot/plot_settings/date_settings.rs
@@ -44,6 +44,7 @@ impl LoadedLogMetadata {
 
 #[derive(PartialEq, Eq, Deserialize, Serialize)]
 pub struct LoadedLogSettings {
+    //
     log_id: usize,
     log_descriptive_name: String,
     pub original_start_date: DateTime<Utc>,
@@ -56,6 +57,8 @@ pub struct LoadedLogSettings {
     show: bool,
     log_metadata: Option<Vec<LoadedLogMetadata>>,
     parse_info: Option<ParseInfo>,
+    marked_for_deletion: bool,
+    is_hovered: bool,
 }
 
 impl LoadedLogSettings {
@@ -84,6 +87,8 @@ impl LoadedLogSettings {
             show: true,
             log_metadata,
             parse_info,
+            marked_for_deletion: false,
+            is_hovered: false,
         }
     }
 
@@ -97,13 +102,14 @@ impl LoadedLogSettings {
 
     pub fn log_label(&self) -> String {
         format!(
-            "#{log_id} {descriptive_name} [{start_date}]",
+            "#{log_id} {descriptive_name}",
             log_id = self.log_id,
             descriptive_name = self.log_descriptive_name,
-            start_date = self.start_date.naive_utc()
+            //start_date = self.start_date.naive_utc()
         )
     }
 
+    /// This is the ID that connects settings to plots
     pub fn log_id(&self) -> usize {
         self.log_id
     }
@@ -122,6 +128,22 @@ impl LoadedLogSettings {
 
     pub fn parse_info(&self) -> Option<ParseInfo> {
         self.parse_info
+    }
+
+    pub fn marked_for_deletion(&self) -> bool {
+        self.marked_for_deletion
+    }
+
+    pub fn marked_for_deletion_mut(&mut self) -> &mut bool {
+        &mut self.marked_for_deletion
+    }
+
+    pub fn cursor_hovering_on(&self) -> bool {
+        self.is_hovered
+    }
+
+    pub fn cursor_hovering_on_mut(&mut self) -> &mut bool {
+        &mut self.is_hovered
     }
 }