+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Added
+
+- Allow deleting individual loaded files without removing everything at once
+
+
 ## [1.3.8]
 
 ### Fix